long amount changed to int

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/logic/IProductService.java
@@ -17,7 +17,7 @@ public interface IProductService {
 
     void deleteProduct(long id) throws NotFoundException;
 
-    long getAmount(long id) throws NotFoundException;
+    int getAmount(long id) throws NotFoundException;
 
-    long modifyAmount(long id, long amount) throws NotFoundException;
+    int modifyAmount(long id, int amount) throws NotFoundException;
 }
