+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/web/bodies/ProductResponse.java
@@ -8,7 +8,7 @@ public class ProductResponse {
     private long id;
     private String name;
     private String description;
-    private long amount;
+    private int amount;
     private String unit;
     private double price;
